Cambios api

+++ Repository/Repositories/UserRepository.cs
@@ -1,7 +1,9 @@
 ﻿using Core.Models;
 using Domain.Interfaces;
-using Repository.Data;
 using Microsoft.EntityFrameworkCore;
+using Repository.Data;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Repository.Repositories
 {
@@ -16,41 +18,44 @@ namespace Repository.Repositories
 
         public async Task<List<User>> GetAllAsync()
         {
-            return await _context.Users.ToListAsync();
+            return await _context.Users.ToListAsync().ConfigureAwait(false);
         }
 
         public async Task<User> GetByIdAsync(int id)
         {
-            return await _context.Users.FindAsync(id);
+            return await _context.Users.FindAsync(id).ConfigureAwait(false);
         }
 
         public async Task<User> CreateAsync(User user)
         {
             _context.Users.Add(user);
-            await _context.SaveChangesAsync();
+            await _context.SaveChangesAsync().ConfigureAwait(false);
             return user;
         }
 
         public async Task<bool> UpdateAsync(User user)
         {
             _context.Entry(user).State = EntityState.Modified;
-            return await _context.SaveChangesAsync() > 0;
+            return await _context.SaveChangesAsync().ConfigureAwait(false) > 0;
         }
 
         public async Task<bool> DeleteAsync(int id)
         {
-            var user = await _context.Users.FindAsync(id);
+            var userExists = await _context.Users.AnyAsync(u => u.Id == id).ConfigureAwait(false);
+            if (!userExists)
+                return false;
+
+            var user = await _context.Users.FindAsync(id).ConfigureAwait(false);
             if (user == null)
                 return false;
 
             _context.Users.Remove(user);
-            return await _context.SaveChangesAsync() > 0;
+            return await _context.SaveChangesAsync().ConfigureAwait(false) > 0;
         }
 
         public async Task<User> GetUserByUsernameAsync(string username)
         {
-            return await _context.Users.FirstOrDefaultAsync(u => u.Username == username);
+            return await _context.Users.FirstOrDefaultAsync(u => u.Username == username).ConfigureAwait(false);
         }
-
     }
 }
